Renaming packages to avoid split them accross multiple modules
Related to #44742 and #44741
