--- extensions/spring-web/resteasy-classic/tests/src/test/java/io/quarkus/spring/web/test/SimpleSpringControllerTest.java
@@ -1,22 +1,0 @@
-package io.quarkus.spring.web.test;
-
-import static io.restassured.RestAssured.when;
-import static org.hamcrest.Matchers.is;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import io.quarkus.test.QuarkusUnitTest;
-
-public class SimpleSpringControllerTest {
-    @RegisterExtension
-    static QuarkusUnitTest runner = new QuarkusUnitTest()
-            .withApplicationRoot((jar) -> jar
-                    .addClasses(SimpleSpringController.class));
-
-    @Test
-    public void testRootResource() {
-        when().get("/simple").then().body(is("hello"));
-    }
-
-}
