+++ extensions/spring-web/resteasy-reactive/tests/src/test/java/io/quarkus/spring/web/resteasy/reactive/test/RequiredFalseTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.spring.web.test;
+package io.quarkus.spring.web.resteasy.reactive.test;
 
 import static io.restassured.RestAssured.when;
 import static org.hamcrest.Matchers.is;
